[First Commit]
Website Skeleton

This is the website skeleton, use this for base site and working. It has all basic functions of a website and already uses django. This does not need XAMMP or WAMP to work python and django can do localhost alone (this might change when we connect the DB).

For all members please try to run this on your own machine (optional). Use the anaconda virtual environment before running ( pyton 3.6 and django 3.1)

NOTES :
 - No connection to database yet (need to change the default DB to MariaDB)
 - Please branch before changing anything.